Show `IsEquivalentTo` and `IsJsonSerializable` directly
